move function to avoid circular dependency
