+++ py/gfa_reduce/analysis/dm.py
@@ -5,42 +5,44 @@ import copy
 from astropy.table import Table
 import time
 import desimeter
-import gfa_reduce.io as io
+from desiutil.log import get_logger
+
+from ..common import retrieve_git_rev
 
 def _check_header_cards(header):
-
+    log = get_logger()
     keywords = ['TARGTRA', 'TARGTDEC', 'FOCUS', 'ADC1PHI', 'ADC2PHI']
 
     for kw in keywords:
         if (kw not in header) or (header[kw] is None) or (type(header[kw]).__name__ == 'Undefined'):
-            print('SKIPPING DESIMETER FIELD MODEL DUE TO MISSING METADATA')
+            log.warning('SKIPPING DESIMETER FIELD MODEL DUE TO MISSING METADATA')
             return False
 
     return True
 
 def fit_dm_fieldmodel(header, ccds, _catalog, check_header_cards=False):
-
+    log = get_logger()
     if check_header_cards:
         good = _check_header_cards(header)
         if not good:
             return None
 
     t0 = time.time()
-    
+
     fm = FieldModel()
 
     catalog = copy.deepcopy(_catalog)
 
     catalog = Table(catalog)
-    
+
     catalog = catalog[catalog['ang_sep_deg'] < 2.0/3600.0]
 
     if len(np.unique(catalog['extname'])) < 2:
-        print('not enough sources with Gaia counterparts for a FieldModel')
+        log.warning('not enough sources with Gaia counterparts for a FieldModel')
         return None
-        
+
     good_ccds = ccds[ccds['contrast'] > 2]['extname']
-    
+
     # restrict to guide cameras with (contrast > 2)
     # if < 2 such guide cameras, then give up
 
@@ -48,13 +50,13 @@ def fit_dm_fieldmodel(header, ccds, _catalog, check_header_cards=False):
     # (contrast > 2) guide cameras
 
     keep = [(_extname in good_ccds) for _extname in catalog['extname']]
-    
+
     # give up if fewer than some minimum number of stars ??
 
     catalog = catalog[keep]
 
     if len(np.unique(catalog['extname'])) < 2:
-        print('not enough sources with Gaia counterparts for a FieldModel')
+        log.warning('not enough sources with Gaia counterparts for a FieldModel')
         return None
 
     fm.ra  = header['TARGTRA'] # other option is SKYRA
@@ -69,8 +71,8 @@ def fit_dm_fieldmodel(header, ccds, _catalog, check_header_cards=False):
 
     for kw in ['ADC1PHI', 'ADC2PHI']:
         if kw not in header:
-            print('WARNING: ' + kw + ' missing !!! assuming ' + kw + ' = 0.0')
-    
+            log.warning(kw + ' missing !!! assuming ' + kw + ' = 0.0')
+
     fm.mjd = np.mean(catalog['mjd_obs'])
     fm.lst = mjd2lst(fm.mjd)
 
@@ -81,8 +83,8 @@ def fit_dm_fieldmodel(header, ccds, _catalog, check_header_cards=False):
 
     # any other metadata that I want to record?
     fm.n_cameras = len(np.unique(catalog['extname']))
-    fm.desimeter_gitrev = io.retrieve_git_rev(fname=desimeter.__file__)
+    fm.desimeter_gitrev = retrieve_git_rev(fname=desimeter.__file__)
 
     dt = time.time()-t0
-    print('total time taken to derive desimeter FieldModel : ', dt, ' seconds')
+    log.info('total time taken to derive desimeter FieldModel : ', dt, ' seconds')
     return fm
